remove chunk in sync
